+++ packages/renderer/src/get-video-threads-flag.ts
@@ -1,5 +1,5 @@
+import {getAvailableMemory} from './get-available-memory';
 import {getCpuCount} from './get-cpu-count';
-import {getAvailableMemory} from './get-free-memory';
 const MEMORY_USAGE_PER_THREAD = 400_000_000; // 400MB
 const RESERVED_MEMORY = 2_000_000_000;
